+++ man-roxygen/papers-probability.R
@@ -1,0 +1,15 @@
+#' @references
+#'
+#' **Data originated from:**
+#'
+#' - Heller, J., & Wickelmaier, F. (2013).
+#'   Minimum discrepancy estimation in probabilistic knowledge structures.
+#'   Electronic Notes in Discrete Mathematics, 42, 49–56.
+#'   \doi{10.1016/j.endm.2013.05.145}
+#'
+#' **Data used in:**
+#'
+#' - Chen, Yinghan, Liu, Y., Culpepper, S. A., & Chen, Y. (2021).
+#'   Inferring the number of attributes for the exploratory DINA model.
+#'   Psychometrika, 86(1), 30–64. \doi{10.1007/s11336-021-09750-9}
+